chore: remove unused publishAgoricNames
